refactor(mapKeys): rename for consistency, kebab-case
